# src/OverEngineeredToDoList.WebApp/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "OverEngineeredToDoList",
    "description": "",
    "termsOfService": "https://example.com/terms",
    "contact": {
      "name": "",
      "email": ""
    },
    "license": {
      "name": "Use under MIT",
      "url": "https://opensource.org/licenses/MIT"
    },
    "version": "v1"
  },
  "paths": {
    "/api/ToDo/{toDoId}": {
      "get": {
        "tags": [
          "ToDo"
        ],
        "summary": "Get ToDo by id.",
        "description": "Get ToDo by id.",
        "operationId": "getToDoById",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "toDoId",
            "required": true,
            "type": "string",
            "format": "uuid"
          }
        ],
        "responses": {
          "404": {
            "description": "Not Found",
            "schema": {
              "type": "string"
            }
          },
          "500": {
            "description": "Server Error"
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          },
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/GetToDoByIdResponse"
            }
          }
        }
      },
      "delete": {
        "tags": [
          "ToDo"
        ],
        "summary": "Delete ToDo.",
        "description": "Delete ToDo.",
        "operationId": "removeToDo",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "toDoId",
            "required": true,
            "type": "string",
            "format": "uuid"
          }
        ],
        "responses": {
          "500": {
            "description": "Server Error"
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          },
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/RemoveToDoResponse"
            }
          }
        }
      }
    },
    "/api/ToDo": {
      "get": {
        "tags": [
          "ToDo"
        ],
        "summary": "Get ToDos.",
        "description": "Get ToDos.",
        "operationId": "getToDos",
        "produces": [
          "application/json"
        ],
        "responses": {
          "500": {
            "description": "Server Error"
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          },
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/GetToDosResponse"
            }
          }
        }
      },
      "post": {
        "tags": [
          "ToDo"
        ],
        "summary": "Create ToDo.",
        "description": "Create ToDo.",
        "operationId": "createToDo",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/CreateToDoRequest"
            }
          }
        ],
        "responses": {
          "500": {
            "description": "Server Error"
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          },
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/CreateToDoResponse"
            }
          }
        }
      },
      "put": {
        "tags": [
          "ToDo"
        ],
        "summary": "Update ToDo.",
        "description": "Update ToDo.",
        "operationId": "updateToDo",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/UpdateToDoRequest"
            }
          }
        ],
        "responses": {
          "500": {
            "description": "Server Error"
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          },
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/UpdateToDoResponse"
            }
          }
        }
      }
    },
    "/api/ToDo/page/{pageSize}/{index}": {
      "get": {
        "tags": [
          "ToDo"
        ],
        "summary": "Get ToDo Page.",
        "description": "Get ToDo Page.",
        "operationId": "getToDosPage",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "pageSize",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "path",
            "name": "index",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "500": {
            "description": "Server Error"
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          },
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/GetToDosPageResponse"
            }
          }
        }
      }
    }
  },
  "definitions": {
    "CreateToDoRequest": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        }
      }
    },
    "CreateToDoResponse": {
      "type": "object",
      "properties": {
        "toDo": {
          "$ref": "#/definitions/ToDoDto"
        },
        "validationErrors": {
          "type": "array",
          "items": {
            "type": "string"
          }
        }
      }
    },
    "GetToDoByIdResponse": {
      "type": "object",
      "properties": {
        "toDo": {
          "$ref": "#/definitions/ToDoDto"
        },
        "validationErrors": {
          "type": "array",
          "items": {
            "type": "string"
          }
        }
      }
    },
    "GetToDosPageResponse": {
      "type": "object",
      "properties": {
        "length": {
          "format": "int32",
          "type": "integer"
        },
        "entities": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/ToDoDto"
          }
        },
        "validationErrors": {
          "type": "array",
          "items": {
            "type": "string"
          }
        }
      }
    },
    "GetToDosResponse": {
      "type": "object",
      "properties": {
        "toDos": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/ToDoDto"
          }
        },
        "validationErrors": {
          "type": "array",
          "items": {
            "type": "string"
          }
        }
      }
    },
    "ProblemDetails": {
      "type": "object",
      "properties": {
        "type": {
          "type": "string"
        },
        "title": {
          "type": "string"
        },
        "status": {
          "format": "int32",
          "type": "integer"
        },
        "detail": {
          "type": "string"
        },
        "instance": {
          "type": "string"
        }
      },
      "additionalProperties": { }
    },
    "RemoveToDoResponse": {
      "type": "object",
      "properties": {
        "toDo": {
          "$ref": "#/definitions/ToDoDto"
        },
        "validationErrors": {
          "type": "array",
          "items": {
            "type": "string"
          }
        }
      }
    },
    "ToDoDto": {
      "type": "object",
      "properties": {
        "toDoId": {
          "format": "uuid",
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "complete": {
          "type": "boolean"
        }
      }
    },
    "UpdateToDoRequest": {
      "type": "object",
      "properties": {
        "toDoId": {
          "format": "uuid",
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "complete": {
          "type": "boolean"
        }
      }
    },
    "UpdateToDoResponse": {
      "type": "object",
      "properties": {
        "toDo": {
          "$ref": "#/definitions/ToDoDto"
        },
        "validationErrors": {
          "type": "array",
          "items": {
            "type": "string"
          }
        }
      }
    }
  }
}